Move types into Tests.UnwrapQQ.Types
